+++ src/Domain/User/Agenda/Version/DiffVerbalizer.php
@@ -15,6 +15,7 @@ use Proximum\Vimeet\Domain\Model\MeetingSlot;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\Spot;
 use Proximum\Vimeet\Domain\Model\User\Agenda\Version;
+use Proximum\Vimeet\Domain\Repository\Meeting\MessageRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\MeetingSlotRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\SheetRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\SpotRepositoryInterface;
@@ -28,10 +29,11 @@ use Proximum\Vimeet\Domain\User\Agenda\Version\View\MeetingMovedView;
 class DiffVerbalizer
 {
     const TRANSLATION_AGENDA_MODIFIED = 'user.agenda.version.agendaModified';
-    const TRANSLATION_MEETING_ADDED   = 'user.agenda.version.meetingAdded';
+    const TRANSLATION_MEETING_ADDED = 'user.agenda.version.meetingAdded';
     const TRANSLATION_MEETING_DELETED = 'user.agenda.version.meetingDeleted';
-    const TRANSLATION_MEETING_MOVED   = 'user.agenda.version.meetingMoved';
-    const TRANSLATION_DOMAIN          = 'messages';
+    const TRANSLATION_MEETING_MOVED = 'user.agenda.version.meetingMoved';
+    const TRANSLATION_MEETING_CHANGED_WITH_MESSAGE = 'user.agenda.version.meetingChangedWithMessage';
+    const TRANSLATION_DOMAIN = 'messages';
 
     /** @var Sheet[] */
     private $sheets;
@@ -55,62 +57,55 @@ class DiffVerbalizer
     private $spotRepository;
 
     /** @var \IntlDateFormatter */
-    private $dayFormatter = null;
+    private $dayFormatter;
 
     /** @var \IntlDateFormatter */
-    private $hourFormatter = null;
+    private $hourFormatter;
 
     /** @var DiffChecker */
     private $diffChecker;
 
-    /**
-     * @param DiffChecker                    $diffChecker
-     * @param TranslatorInterface            $translator
-     * @param SheetRepositoryInterface       $sheetRepository
-     * @param MeetingSlotRepositoryInterface $meetingSlotRepository
-     * @param SpotRepositoryInterface        $spotRepository
-     */
+    /** @var MessageRepositoryInterface */
+    private $messageRepository;
+
+    /** @var string[] */
+    private $messages;
+
     public function __construct(
         DiffChecker $diffChecker,
         TranslatorInterface $translator,
         SheetRepositoryInterface $sheetRepository,
         MeetingSlotRepositoryInterface $meetingSlotRepository,
-        SpotRepositoryInterface $spotRepository
+        SpotRepositoryInterface $spotRepository,
+        MessageRepositoryInterface $messageRepository
     ) {
         $this->diffChecker = $diffChecker;
         $this->translator = $translator;
         $this->sheetRepository = $sheetRepository;
         $this->meetingSlotRepository = $meetingSlotRepository;
         $this->spotRepository = $spotRepository;
+        $this->messageRepository = $messageRepository;
     }
 
-    /**
-     * @param Version $lastVersion
-     * @param array   $currentVersion
-     * @param string  $locale
-     *
-     * @return string
-     */
     public function verbalizeDiff(Version $lastVersion, array $currentVersion, string $locale): string
     {
         $diffs = [];
 
         $deletedRequests = array_diff_key($lastVersion->getVersion(), $currentVersion);
-        $addedRequests   = array_diff_key($currentVersion, $lastVersion->getVersion());
-        $toCheckChanges  = array_intersect_key($lastVersion->getVersion(), $currentVersion);
+        $addedRequests = array_diff_key($currentVersion, $lastVersion->getVersion());
+        $toCheckChanges = array_intersect_key($lastVersion->getVersion(), $currentVersion);
 
         if (empty($deletedRequests) && empty($addedRequests) && empty($toCheckChanges)) {
             return '';
         }
 
         $userSheets = $this->sheetRepository
-            ->getSheetsByUserAndEvent($lastVersion->getUser(), $lastVersion->getEvent())
-        ;
+            ->getSheetsByUserAndEvent($lastVersion->getUser(), $lastVersion->getEvent());
 
         $deletedRequestViews = $this->getDeletedMeeting($userSheets, $deletedRequests);
         $addedRequestViews = $this->getAddedMeeting($userSheets, $addedRequests);
         $changedViews = $this->getChangedMeeting($userSheets, $toCheckChanges, $currentVersion);
-        $this->getInfo($deletedRequestViews, $addedRequestViews, $changedViews);
+        $this->getInfo($deletedRequestViews, $addedRequestViews, $changedViews, $userSheets);
 
         $this->dayFormatter = new \IntlDateFormatter(
             $locale,
@@ -137,16 +132,28 @@ class DiffVerbalizer
      * @param MeetingDeletedView[] $deletedRequestViews
      * @param string               $locale
      */
-    private function addDeletedSentences(array &$diffs, array &$deletedRequestViews, string $locale)
+    private function addDeletedSentences(array &$diffs, array &$deletedRequestViews, string $locale): void
     {
         foreach ($deletedRequestViews as $deletedRequestView) {
             if (isset($this->sheets[$deletedRequestView->sheetId])) {
-                $diffs[] = $this->translator->trans(
+                $sentence = $this->translator->trans(
                     self::TRANSLATION_MEETING_DELETED,
                     ['%sheetTitle%' => $this->sheets[$deletedRequestView->sheetId]->getTitle()],
                     self::TRANSLATION_DOMAIN,
                     $locale
                 );
+
+                $latestMessage = $this->messages[$deletedRequestView->requestId] ?? null;
+                if ($latestMessage !== null) {
+                    $sentence .= ' ' . $this->translator->trans(
+                            self::TRANSLATION_MEETING_CHANGED_WITH_MESSAGE,
+                            ['%message%' => $latestMessage],
+                            self::TRANSLATION_DOMAIN,
+                            $locale
+                        );
+                }
+
+                $diffs[] = $sentence;
             }
         }
     }
@@ -156,7 +163,7 @@ class DiffVerbalizer
      * @param MeetingAddedView[] $addedRequestViews
      * @param string             $locale
      */
-    private function addAddedSentences(array &$diffs, array &$addedRequestViews, string $locale)
+    private function addAddedSentences(array &$diffs, array &$addedRequestViews, string $locale): void
     {
         foreach ($addedRequestViews as $addedRequestView) {
             if (isset($this->sheets[$addedRequestView->sheetId])
@@ -183,14 +190,14 @@ class DiffVerbalizer
      * @param MeetingMovedView[] $changedMeetings
      * @param string             $locale
      */
-    private function addMovedSentences(array &$diffs, array &$changedMeetings, string $locale)
+    private function addMovedSentences(array &$diffs, array &$changedMeetings, string $locale): void
     {
         foreach ($changedMeetings as $changedMeeting) {
             if (isset($this->sheets[$changedMeeting->sheetId])
                 && isset($this->slots[$changedMeeting->slotId])
                 && isset($this->spots[$changedMeeting->spotId])
             ) {
-                $diffs[] = $this->translator->trans(
+                $sentence = $this->translator->trans(
                     self::TRANSLATION_MEETING_MOVED,
                     [
                         '%sheetTitle%' => $this->sheets[$changedMeeting->sheetId]->getTitle(),
@@ -201,6 +208,18 @@ class DiffVerbalizer
                     self::TRANSLATION_DOMAIN,
                     $locale
                 );
+
+                $latestMessage = $this->messages[$changedMeeting->requestId] ?? null;
+                if ($latestMessage !== null) {
+                    $sentence .= ' ' . $this->translator->trans(
+                            self::TRANSLATION_MEETING_CHANGED_WITH_MESSAGE,
+                            ['%message%' => $latestMessage],
+                            self::TRANSLATION_DOMAIN,
+                            $locale
+                        );
+                }
+
+                $diffs[] = $sentence;
             }
         }
     }
@@ -211,9 +230,14 @@ class DiffVerbalizer
      * @param MeetingDeletedView[] $deletedRequestViews
      * @param MeetingAddedView[]   $addedRequestViews
      * @param MeetingMovedView[]   $changedViews
+     * @param Sheet[]              $userSheets
      */
-    private function getInfo(array $deletedRequestViews, array $addedRequestViews, array $changedViews)
-    {
+    private function getInfo(
+        array $deletedRequestViews,
+        array $addedRequestViews,
+        array $changedViews,
+        array $userSheets
+    ): void {
         $sheets = [];
         $slots = [];
         $spots = [];
@@ -237,6 +261,15 @@ class DiffVerbalizer
         $this->sheets = !empty($sheets) ? $this->sheetRepository->findByIds($sheets) : [];
         $this->slots = !empty($slots) ? $this->meetingSlotRepository->findByIds($slots) : [];
         $this->spots = !empty($spots) ? $this->spotRepository->getSpotsByIds($spots) : [];
+
+        $concernedByMessagesRequestViews = array_merge($deletedRequestViews, $changedViews);
+        $concernedByMessagesRequestIds = array_map(
+            static function ($item) {
+                return $item->requestId;
+            },
+            $concernedByMessagesRequestViews
+        );
+        $this->messages = $this->getLatestMessagesByRequest($concernedByMessagesRequestIds, $userSheets);
     }
 
     /**
@@ -251,7 +284,9 @@ class DiffVerbalizer
 
         if (!empty($deletedRequests)) {
             foreach ($deletedRequests as $key => $request) {
-                $deletedViews[] = new MeetingDeletedView($this->getSheetMet($userSheets, $request));
+                $deletedViews[] = new MeetingDeletedView(
+                    $this->getSheetMet($userSheets, $request), $request['request']
+                );
             }
         }
 
@@ -281,13 +316,6 @@ class DiffVerbalizer
         return $addedViews;
     }
 
-    /**
-     * @param array $userSheets
-     * @param array $intersectVersion
-     * @param array $currentVersion
-     *
-     * @return array
-     */
     private function getChangedMeeting(array &$userSheets, array &$intersectVersion, array &$currentVersion): array
     {
         $changedViews = [];
@@ -298,7 +326,8 @@ class DiffVerbalizer
                     $changedViews[] = new MeetingMovedView(
                         $this->getSheetMet($userSheets, $request),
                         $currentVersion[$key]['slot'],
-                        $currentVersion[$key]['spot']
+                        $currentVersion[$key]['spot'],
+                        $request['request']
                     );
                 }
             } catch (\InvalidArgumentException $exception) {
@@ -320,36 +349,43 @@ class DiffVerbalizer
         return isset($userSheets[$request['fromSheet']]) ? $request['toSheet'] : $request['fromSheet'];
     }
 
-    /**
-     * @param \DateTimeInterface $date
-     *
-     * @return string
-     */
     private function formatDay(\DateTimeInterface $date): string
     {
         return $this->format($this->dayFormatter, $date);
     }
 
-    /**
-     * @param \DateTimeInterface $hour
-     *
-     * @return string
-     */
     private function formatHour(\DateTimeInterface $hour): string
     {
         return $this->format($this->hourFormatter, $hour);
     }
 
-    /**
-     * @param \IntlDateFormatter $dateFormatter
-     * @param \DateTimeInterface $date
-     *
-     * @return string
-     */
     private function format(\IntlDateFormatter $dateFormatter, \DateTimeInterface $date): string
     {
         $dateFormatted = $dateFormatter->format($date);
 
         return \is_bool($dateFormatted) ? '' : $dateFormatted;
+    }
+
+    /**
+     * @param int[]   $requestIds
+     * @param Sheet[] $userSheets
+     *
+     * @return string[]
+     */
+    private function getLatestMessagesByRequest(array $requestIds, array $userSheets): array
+    {
+        // get latest message by request
+        $messages = $this->messageRepository->getUpdateOrDeleteReasonMessageFromRequestIds($requestIds);
+
+        $indexedMessages = [];
+        foreach ($messages as $message) {
+            // ignore message if it's mine
+            if (\in_array($message->getFrom(), $userSheets, true)) {
+                continue;
+            }
+            $indexedMessages[$message->getRequest()->getId()] = $message->getContent();
+        }
+
+        return $indexedMessages;
     }
 }