+++ tests/Application/Command/Meeting/Admin/RemoveMeetingHandlerTest.php
@@ -23,6 +23,7 @@ use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\MeetingSlot;
 use Proximum\Vimeet\Domain\Model\Participant;
 use Proximum\Vimeet\Domain\Model\Spot;
+use Proximum\Vimeet\Domain\Repository\Meeting\RequestRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\MeetingRepositoryInterface;
 use Proximum\Vimeet\Infrastructure\Adapter\DelayedEventDispatcher;
 use Proximum\Vimeet\Tests\Factory\EventFactory;
@@ -57,6 +58,7 @@ class RemoveMeetingHandlerTest extends TestCase
         $meetingRepository = $this->prophesize(MeetingRepositoryInterface::class);
         $translator        = $this->prophesize(TranslatorInterface::class);
         $eventDispatcher   = $this->prophesize(DelayedEventDispatcher::class);
+        $requestDispatcher = $this->prophesize(RequestRepositoryInterface::class);
 
         if (!$meeting->isBlockedSlot()) {
             $meetingRepository->remove($meeting)->shouldBeCalled();
@@ -85,10 +87,13 @@ class RemoveMeetingHandlerTest extends TestCase
         $handler = new RemoveMeetingHandler(
             $meetingRepository->reveal(),
             $translator->reveal(),
-            $eventDispatcher->reveal()
+            $eventDispatcher->reveal(),
+            $requestDispatcher->reveal()
         );
 
         $handler->handle(new RemoveMeeting($meeting, $admin));
+
+        $this->assertNull($meeting->getRequest()->getUpdateOrDeleteReasonMessage());
     }
 
     private function getMeeting($blockedSlot = false)
@@ -100,6 +105,7 @@ class RemoveMeetingHandlerTest extends TestCase
         $sheet2 = SheetFactory::create($event);
         $this->participant1 = $this->prophesize(Participant::class);
         $this->participant2 = $this->prophesize(Participant::class);
+        $message = $this->prophesize(Meeting\Message::class);
 
         $request = new Meeting\Request(
             $sheet1,
@@ -110,6 +116,7 @@ class RemoveMeetingHandlerTest extends TestCase
             $user,
             $event
         );
+        $request->setUpdateOrDeleteReasonMessage($message->reveal());
         $slot = new MeetingSlot($event, $dateTime, $dateTime, $blockedSlot);
         $spot = new Spot('ref', $event, 100, 200, 150, true);