+++ tests/Application/Command/Meeting/Event/RemoveHandlerTest.php
@@ -18,6 +18,7 @@ use Proximum\Vimeet\Application\Command\Meeting\Event\RemoveHandler;
 use Proximum\Vimeet\Domain\Meeting\CanRemoveMeeting;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Repository\Meeting\RequestRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\MeetingRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\Meeting\MessageRepositoryInterface;
 
@@ -33,6 +34,11 @@ class RemoveHandlerTest extends TestCase
         $sheet,
         $meeting;
 
+    /** @var ObjectProphecy|RequestRepositoryInterface */
+    private $requestRepository;
+    /** @var ObjectProphecy|Meeting\Request */
+    private $request;
+
     public function setUp()
     {
         $this->canRemoveMeeting = $this->prophesize(CanRemoveMeeting::class);
@@ -42,6 +48,9 @@ class RemoveHandlerTest extends TestCase
         $this->datetime = new \DateTime();
         $this->sheet = $this->prophesize(Sheet::class);
         $this->meeting = $this->prophesize(Meeting::class);
+        $this->requestRepository = $this->prophesize(RequestRepositoryInterface::class);
+        $this->request = $this->prophesize(Meeting\Request::class);
+        $this->meeting->getRequest()->willReturn($this->request->reveal());
     }
 
     /**
@@ -55,13 +64,15 @@ class RemoveHandlerTest extends TestCase
 
         $this->messageRepository->add(Argument::any())->shouldNotBeCalled();
         $this->meetingRepository->set(Argument::any())->shouldNotBeCalled();
+        $this->requestRepository->set(Argument::any())->shouldNotBeCalled();
 
         $handler = new RemoveHandler(
             $this->messageRepository->reveal(),
             $this->meetingRepository->reveal(),
             $this->delayedEventDispatcher->reveal(),
             $this->canRemoveMeeting->reveal(),
-            $this->datetime
+            $this->datetime,
+            $this->requestRepository->reveal()
         );
         $handler->handle(new Remove($this->sheet->reveal(), $this->meeting->reveal()));
     }
@@ -79,13 +90,15 @@ class RemoveHandlerTest extends TestCase
 
         $this->messageRepository->add(Argument::any())->shouldNotBeCalled();
         $this->meetingRepository->remove($this->meeting->reveal())->shouldBeCalled()->willThrow(new \Exception());
+        $this->requestRepository->set(Argument::any())->shouldNotBeCalled();
 
         $handler = new RemoveHandler(
             $this->messageRepository->reveal(),
             $this->meetingRepository->reveal(),
             $this->delayedEventDispatcher->reveal(),
             $this->canRemoveMeeting->reveal(),
-            $this->datetime
+            $this->datetime,
+            $this->requestRepository->reveal()
         );
 
         $remove = new Remove($this->sheet->reveal(), $this->meeting->reveal());
@@ -115,7 +128,8 @@ class RemoveHandlerTest extends TestCase
             $this->meetingRepository->reveal(),
             $this->delayedEventDispatcher->reveal(),
             $this->canRemoveMeeting->reveal(),
-            $this->datetime
+            $this->datetime,
+            $this->requestRepository->reveal()
         );
 
         $remove = new Remove($this->sheet->reveal(), $this->meeting->reveal());
@@ -131,25 +145,26 @@ class RemoveHandlerTest extends TestCase
         $this->meeting->hasSheet($this->sheet->reveal())->shouldBeCalled()->willReturn(true);
         $this->meeting->getFromSheet()->shouldBeCalled();
         $this->meeting->getToSheet()->shouldBeCalled();
-        $this->meeting->getAllParticipants()->shouldBeCalled()->willReturn([]);
+        $this->meeting->getAllParticipants()->shouldBeCalled()->willReturn([]);;
+
+        $this->request->setUpdateOrDeleteReasonMessage(Argument::type(Meeting\Message::class))->shouldBeCalled();
 
         $this->canRemoveMeeting->isSatisfiedBy($this->sheet->reveal())
             ->shouldBeCalled()
             ->willReturn(true);
 
-        $request = $this->prophesize(Meeting\Request::class);
-        $this->meeting->getRequest()->shouldBeCalled()->willReturn($request->reveal());
-
-        $message = new Meeting\Message($request->reveal(), $this->sheet->reveal(), 'content', $this->datetime);
+        $message = new Meeting\Message($this->request->reveal(), $this->sheet->reveal(), 'content', $this->datetime);
         $this->messageRepository->add($message)->shouldBeCalled();
         $this->meetingRepository->remove($this->meeting->reveal())->shouldBeCalled();
+        $this->requestRepository->set($this->request->reveal())->shouldBeCalled();
 
         $handler = new RemoveHandler(
             $this->messageRepository->reveal(),
             $this->meetingRepository->reveal(),
             $this->delayedEventDispatcher->reveal(),
             $this->canRemoveMeeting->reveal(),
-            $this->datetime
+            $this->datetime,
+            $this->requestRepository->reveal()
         );
 
         $remove = new Remove($this->sheet->reveal(), $this->meeting->reveal());