+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- VIMEET-1904 - Add message to meeting diff
 - VIMEET-1967 - Package can be required by type
 - VIMEET-1968 - Payment can be required by type
 - VIMEET-1965 - Must validate package